requirements.txt
# application dependencies
django==4.1
django-admin==2.0.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
psycopg2==2.9.5

# test dependencies
pytest-cov==4.0.0
pytest-django==4.5.2
requests==2.28.2
